requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
anyio==3.6.2
appnope==0.1.3
arrow==1.2.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.12.2
binaryornot==0.4.4
blessed==1.20.0
cachetools==5.3.0
certifi==2022.12.7
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
colorlog==6.7.0
comm==0.1.3
contourpy==1.0.7
cookiecutter==2.1.1
croniter==1.3.10
cycler==0.11.0
dateutils==0.6.12
debugpy==1.6.7
decorator==5.1.1
deepdiff==6.3.0
dnspython==2.3.0
email-validator==1.3.1
executing==1.2.0
fastapi==0.88.0
filelock==3.11.0
fonttools==4.39.3
frozenlist==1.3.3
fsspec==2023.4.0
google-auth==2.17.2
google-auth-oauthlib==1.0.0
grpcio==1.53.0
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
hydra-colorlog==1.2.0
hydra-core==1.3.2
idna==3.4
inquirer==3.1.3
install==1.3.5
ipykernel==6.22.0
ipython==8.12.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jinja2-time==0.2.0
joblib==1.2.0
jupyter_client==8.1.0
jupyter_core==5.3.0
kiwisolver==1.4.4
lightning==2.0.1
lightning-cloud==0.5.32
lightning-utilities==0.8.0
llvmlite==0.39.1
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
nest-asyncio==1.5.6
networkx==3.1
numba==0.56.4
numpy==1.23.5
oauthlib==3.2.2
omegaconf==2.3.0
ordered-set==4.1.0
orjson==3.8.9
packaging==23.0
pandas==2.0.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.2.0
prompt-toolkit==3.0.38
protobuf==4.22.1
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.7
Pygments==2.14.0
PyJWT==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-editor==1.0.4
python-multipart==0.0.6
python-slugify==8.0.1
pytorch-lightning==2.0.1
pytz==2023.3
PyYAML==6.0
pyzmq==25.0.2
readchar==4.0.5
requests==2.28.2
requests-oauthlib==1.3.1
rfc3986==1.5.0
rich==13.3.3
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
# scikit-learn==0.0.post1
sniffio==1.3.0
soupsieve==2.4
stack-data==0.6.2
starlette==0.22.0
starsessions==1.3.0
sympy==1.11.1
tensorboard==2.12.1
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
text-unidecode==1.3
threadpoolctl==3.1.0
torch==2.0.0
torchaudio==2.0.1
torchmetrics==0.11.4
torchvision==0.15.1
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2023.3
ujson==5.7.0
urllib3==1.26.15
uvicorn==0.21.1
uvloop==0.17.0
watchfiles==0.19.0
wcwidth==0.2.6
websocket-client==1.5.1
websockets==11.0.1
Werkzeug==2.2.3
yarl==1.8.2
